requirements.txt
django
django-localflavor
Pillow
